Use IP endpoint lock for multicast group state
The various sub-states of an IP endpoint state (address state, group
state, and NDP state (for IPv6)) are coupled in such a way that using
separate locks for each sub-state is complicated.

Instead of using different locks for each sub-state, use a single global
lock for the whole IP endpoint state.

Startblock:
  has LGTM from peterjohnston
  and then
  add reviewer brunodalbo
PiperOrigin-RevId: 345590869